Module MH.Make

Metropolis-Hastings functor.

Parameters

module X : MH_parameters

Signature

val mcmc : verbosity:[ `Silent | `Progress | `Trace ] -> initial:X.t -> burn_in:int -> X.t Stats.gen

The mcmc function produces a generative probability. Be warned that the samples obtained from mcmc are _not_ independent. The guarantee is that (if the proposal and weight are reasonable) the empirical measure will converge in law to the target measure.