Module StaTz.Series

type StaTz__Result.error +=

| Grid_dimensions_mismatch : 'a Grid.t -> StaTz__Result.error

type ('grid, 'elt, 'data) t

Type of series discretized over a grid of type 'grid and with elements of type 'elt. 'data is the type of the underlying representation.

type 'a elt

Type of admissible elements for the series.

val float_elt : float elt

Admissible elements.

val complex_elt : Stdlib.Complex.t elt

type ('elt, 'data) representation

Allowed underlying representation for the time series data.

val dense_float64 : (float, Numerics.Float64.Vec.t) representation

val dense_complex64 : (Stdlib.Complex.t, Numerics.Complex64.Vec.t) representation

val fun_index : ('e, int -> 'e) representation

val fun_time : ('e, float -> 'e) representation

val grid : ('g, 'e, 'a) t -> 'g Grid.t

Get grid

val data : ('g, 'e, 'a) t -> 'a

Get data

val length : ('g, 'e, 'a) t -> int

Get length of data

val make : ('e, 'a) representation -> 'g Grid.t -> 'a -> ('g, 'e, 'a) t

Make signal from data and grid

val discretise : ('e, 'a) representation -> (float -> 'e) -> 'g Grid.t -> ('g, 'e, 'a) t

Discretise a function over a regular grid.

val eval_t : ('g, 'e, 'a) t -> float -> 'e

Evaluate time series at given time. This is logarithmic in the length of the data in case of a dense representation.

val eval_i : ('g, 'e, 'a) t -> int -> 'e

Evaluate time series at given index. This is a constant-time operation.

val map : ('e2, 'a2) representation -> ('e1 -> 'e2) -> ('g, 'e1, 'a1) t -> ('g, 'e2, 'a2) t

Map a function over a series.

val lift2 : ('a -> 'a -> 'a) -> ('g, 'e, 'a) t -> ('g, 'e, 'a) t -> (('g, 'e, 'a) t, StaTz__Result.error list) Stdlib.result

Lifts a binary operator on data representations to series. Fails if the grids are different.